Fix crash when unable to get booking data
The program would previously crash due to a KeyError when trying to
access the data in the response from the timetree API. If no data was
returned, a KeyError would occur and cause the program to crash. This
commit puts the attempt to retreive the data throught a try except to
appropriately handle the case when no data was returned, and continues
to generate the document with available data.